+++ cmplibrary/src/test/java/com/sourcepoint/gdpr_cmplibrary/exception/LoggerImplTest.kt
@@ -1,5 +1,6 @@
 package com.sourcepoint.gdpr_cmplibrary.exception
 
+import com.example.gdpr_cmplibrary.BuildConfig
 import com.sourcepoint.gdpr_cmplibrary.assertEquals
 import com.sourcepoint.gdpr_cmplibrary.readText
 import io.mockk.MockKAnnotations
@@ -31,18 +32,23 @@ class LoggerImplTest {
         val json = json(CodeList.INVALID_RESPONSE_WEB_MESSAGE)
         every { messageManager.build(ex) } returns json
 
-        val sut = createLogger(client, messageManager, "https://myserver.com/")
+        val cb: (String, String) -> Unit = { _, _ -> }
+        val sut = createLogger4Testing(cb, cb, cb, client, messageManager, "https://myserver.com/")
         sut.error(ex)
 
-        val slot = slot<Request>()
-        verify(exactly = 1) { client.newCall(capture(slot)) }
-
-        slot.captured.run {
-            readText().assertEquals(json)
-            url.toString().assertEquals("https://myserver.com/")
-            method.assertEquals("POST")
+        /** We have 2 different implementation for Debug and Release */
+        when (BuildConfig.DEBUG) {
+            true -> verify(exactly = 0) { client.newCall(any()) }
+            false -> {
+                val slot = slot<Request>()
+                verify(exactly = 1) { client.newCall(capture(slot)) }
+                slot.captured.run {
+                    readText().assertEquals(json)
+                    url.toString().assertEquals("https://myserver.com/")
+                    method.assertEquals("POST")
+                }
+            }
         }
-
     }
 
     @Test
@@ -51,18 +57,23 @@ class LoggerImplTest {
         val json = json(ExceptionCodes("custom_code"))
         every { messageManager.build(ex) } returns json
 
-        val sut = createLogger(client, messageManager, "https://myserver.com/")
+        val cb: (String, String) -> Unit = { _, _ -> }
+        val sut = createLogger4Testing(cb, cb, cb, client, messageManager, "https://myserver.com/")
         sut.error(ex)
 
-        val slot = slot<Request>()
-        verify(exactly = 1) { client.newCall(capture(slot)) }
-
-        slot.captured.run {
-            readText().assertEquals(json)
-            url.toString().assertEquals("https://myserver.com/")
-            method.assertEquals("POST")
+        /** We have 2 different implementation for Debug and Release */
+        when (BuildConfig.DEBUG) {
+            true -> verify(exactly = 0) { client.newCall(any()) }
+            false -> {
+                val slot = slot<Request>()
+                verify(exactly = 1) { client.newCall(capture(slot)) }
+                slot.captured.run {
+                    readText().assertEquals(json)
+                    url.toString().assertEquals("https://myserver.com/")
+                    method.assertEquals("POST")
+                }
+            }
         }
-
     }
 
     private fun json(errorCode: ExceptionCodes) = """