+++ samples/metaapp/src/androidTest/java/com/sourcepointmeta/metaapp/MetaAppTestsK.kt
@@ -44,7 +44,6 @@ import com.sourcepointmeta.metaapp.MetaAppTestCases.Companion.tapSaveAndExitOnWe
 import com.sourcepointmeta.metaapp.MetaAppTestCases.Companion.tapShowOption
 import com.sourcepointmeta.metaapp.TestData.*
 import com.sourcepointmeta.metaapp.ui.SplashScreenActivity
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import org.junit.After
 import org.junit.Test
@@ -84,21 +83,21 @@ class MetaAppTestsK {
 
         scenario = launchActivity()
 
-        wr { tapOnAddProperty() }                           //  tapOnAddProperty();
-        wr { addNativeMessagePropertyDetails() }            //  addNativeMessagePropertyDetails();
-        wr { tapOnSave() }                                  //  tapOnSave();
-        wr { checkNativeMessageDisplayed() }                //  Assert.assertTrue(checkNativeMessageDisplayed());
-        wr { tapRejectAll() }                               //  chooseNativeMessageAction(R.id.RejectAll);
-        wr { checkForPropertyInfoScreen() }                 //  Assert.assertTrue(checkFor(PROPERTY_INFO_SCREEN));
-        wr { loadPrivacyManagerDirect() }                   //  loadPrivacyManagerDirect();
-        wr { checkWebViewDisplayedForPrivacyManager() }     //  Assert.assertTrue(checkWebViewDisplayedFor(PRIVACY_MANAGER));
-        wr { selectNativeMessageConsentList() }             //  Assert.assertFalse(checkConsentsAsSelected(NATIVE_MESSAGE_CONSENT_LIST));
-        wr { selectPartialConsentList() }                   //  selectConsents(PARTIAL_CONSENT_LIST);
-        wr { tapSaveAndExitOnWebView() }                    //  chooseAction(PM_SAVE_AND_EXIT);
-        wr { checkForPropertyInfoScreen() }                 //  Assert.assertTrue(checkFor(PROPERTY_INFO_SCREEN));
-        wr { loadPrivacyManagerDirect() }                   //  loadPrivacyManagerDirect();
-        wr { checkWebViewDisplayedForPrivacyManager() }     //  Assert.assertTrue(checkWebViewDisplayedFor(PRIVACY_MANAGER));
-        wr { selectPartialConsentList() }                   //  Assert.assertTrue(checkConsentsAsSelected(PARTIAL_CONSENT_LIST));
+        wr { tapOnAddProperty() }                               //  tapOnAddProperty();
+        wr { addNativeMessagePropertyDetails() }                //  addNativeMessagePropertyDetails();
+        wr { tapOnSave() }                                      //  tapOnSave();
+        wr { checkNativeMessageDisplayed() }                    //  Assert.assertTrue(checkNativeMessageDisplayed());
+        wr { tapRejectAll() }                                   //  chooseNativeMessageAction(R.id.RejectAll);
+        wr { checkForPropertyInfoScreen() }                     //  Assert.assertTrue(checkFor(PROPERTY_INFO_SCREEN));
+        wr { loadPrivacyManagerDirect() }                       //  loadPrivacyManagerDirect();
+        wr { checkWebViewDisplayedForPrivacyManager() }         //  Assert.assertTrue(checkWebViewDisplayedFor(PRIVACY_MANAGER));
+        wr { selectNativeMessageConsentList() }                 //  Assert.assertFalse(checkConsentsAsSelected(NATIVE_MESSAGE_CONSENT_LIST));
+        wr { selectPartialConsentList() }                       //  selectConsents(PARTIAL_CONSENT_LIST);
+        wr { tapSaveAndExitOnWebView() }                        //  chooseAction(PM_SAVE_AND_EXIT);
+        wr { checkForPropertyInfoScreen() }                     //  Assert.assertTrue(checkFor(PROPERTY_INFO_SCREEN));
+        wr { loadPrivacyManagerDirect() }                       //  loadPrivacyManagerDirect();
+        wr { checkWebViewDisplayedForPrivacyManager() }         //  Assert.assertTrue(checkWebViewDisplayedFor(PRIVACY_MANAGER));
+        wr { selectPartialConsentList() }                       //  Assert.assertTrue(checkConsentsAsSelected(PARTIAL_CONSENT_LIST));
     }
 
     @Test
@@ -374,7 +373,7 @@ class MetaAppTestsK {
         wr { checkWebViewDisplayedForMessage() }                    //        Assert.assertTrue(checkWebViewDisplayedFor(MESSAGE));
         wr { tapAcceptAllOnWebView() }                              //        chooseAction(ACCEPT_ALL);
         wr { checkForConsentsAreDisplayed() }                       //        Assert.assertTrue(checkFor(CONSENTS_ARE_DISPLAYED));
-        navigateBackToListView()                                    //        navigateBackToListView();
+        wr { navigateBackToListView() }                             //        navigateBackToListView();
         wr { swipeAndChooseEditActionUniqueAuth() }                 //        swipeAndChooseAction(EDIT_ACTION, UNIQUE_AUTHENTICATION);
         wr { checkWebViewDoesNotDisplayTheMessage() }               //        Assert.assertFalse(checkWebViewDisplayedFor(MESSAGE));
         wr { checkForConsentsAreDisplayed() }                       //        Assert.assertTrue(checkFor(CONSENTS_ARE_DISPLAYED));