+++ ui-test-util/src/main/java/com/example/uitestutil/EspressoUtils.kt
@@ -27,7 +27,7 @@ fun isDisplayedAllOfByResId(
 
 fun isDisplayedByResIdByText(
     @IdRes resId: Int,
-    text : String
+    text: String
 ) {
     onView(
         allOf(
@@ -77,7 +77,7 @@ fun performClickById(
 
 @Throws(Throwable::class)
 fun performClickByText(
-    text : String
+    text: String
 ) {
     onView(
         allOf(
@@ -98,10 +98,10 @@ fun performClickContent(
     ).perform(ViewActions.click())
 }
 
-fun performSpinnerItemSelection(@IdRes resId: Int , contentDescription: String){
-    performClickById(resId )
+fun performSpinnerItemSelection(@IdRes resId: Int, contentDescription: String) {
+    performClickById(resId)
     Espresso.onData(CoreMatchers.allOf(CoreMatchers.`is`(CoreMatchers.instanceOf(String::class.java)), CoreMatchers.`is`(contentDescription)))
-            .perform(ViewActions.click())
+        .perform(ViewActions.click())
 }
 
 @Throws(Throwable::class)
@@ -144,7 +144,7 @@ fun checkWebViewHasText(text: String) {
 }
 
 @Throws(Throwable::class)
-fun checkElementWithText(id : String, expected: String) {
+fun checkElementWithText(id: String, expected: String) {
     onWebView()
         .withElement(findElement(Locator.ID, id))
         .check(webMatches(getText(), containsString(expected)));
@@ -167,20 +167,20 @@ fun performClickOnWebViewByClass(classValue: String) {
 }
 
 @Throws(Throwable::class)
-fun checkConsentState(consent: String, selected : Boolean) {
+fun checkConsentState(consent: String, selected: Boolean) {
     onWebView()
         .withElement(findElement(Locator.XPATH, "//label[@aria-label='$consent']"))
         .withElement(findElement(Locator.XPATH, "//label[@aria-checked='$selected']"))
 }
 
 @Throws(Throwable::class)
-fun checkPMTabSelected( expected : String){
+fun checkPMTabSelected(expected: String) {
     onWebView()
         .withElement(findElement(Locator.XPATH, "//div[contains(@class, 'pm-tab active') and text()='$expected']"))
 }
 
 @Throws(Throwable::class)
-fun performClickPMTabSelected( expected : String){
+fun performClickPMTabSelected(expected: String) {
     onWebView()
         .withElement(findElement(Locator.XPATH, "//div[contains(@class, 'pm-tab') and text()='$expected']"))
         .perform(webScrollIntoView())
@@ -188,7 +188,7 @@ fun performClickPMTabSelected( expected : String){
 }
 
 @Throws(Throwable::class)
-fun setCheckBoxTrue(property : String){
+fun setCheckBoxTrue(property: String) {
     onWebView()
         .withElement(findElement(Locator.XPATH, "//label[@aria-label='$property']/span[@class='on']"))
         .perform(webScrollIntoView())
@@ -234,5 +234,7 @@ fun checkWebViewDoesNotHasText(text: String) {
         throw InvalidParameterException("""
             The current view with text {$text} is displayed. 
         """.trimIndent())
-    } catch (e: Exception) { /** This is the success case */ }
+    } catch (e: Exception) {
+        /** This is the success case */
+    }
 }